Allow event triggers to be removed with the callback stored.
Test implemented demonstrating the bug.
.on('evt', fnt) and if that fnt implements off('evt') it would not work.  This has been fixed.

For most use cases using once is enough, but this does not cover every possible way to use on and off for event subscriptions.

A previous pull request did not modify the source and only the build output.
737 /src/audio5.js
957 /src/audio5.js

Without the source being updated, when the grunt task is ran, the changes are reverted.